% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{build_date}
\alias{build_date}
\title{Combine year, month and day numeric features into a date.}
\usage{
build_date(y_raw,m_raw,d_raw)
}
\arguments{
\item{y_raw}{year numeric}

\item{m_raw}{month numeric}

\item{d_raw}{date numeric}
}
\value{
Date
}
\description{
Take 3 numeric as input: year, mounth and day (of the month).
Year must be a valid 4 digit year. If month is NA, month is set to 01 (January),
if day is NA, day of the month is set to 1.
}
\details{
If year is null return NA
If month is null set moth to january
If day is null set day to 1
}
