% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{clean_location}
\alias{clean_location}
\title{Format the LOCATION_NAME colomn.}
\usage{
clean_location(location)
}
\arguments{
\item{location}{character string}
}
\value{
chahracter string
}
\description{
Feature LOCATION_NAME is encoded as COUNTRY: Location. All set to upper case.
Remove the country and capitalize the rest.
}
