% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{eq_build_date}
\alias{eq_build_date}
\title{Building dates form NOAA raw data}
\usage{
eq_build_date(data)
}
\arguments{
\item{data}{raw NOAA data}
}
\value{
Date
}
\description{
Take 3 numeric as input: year, mounth and day (of the month).
Year must be a valid 4 digit year. If month is NA, month is set to 01 (January),
if day is NA, day of the month is set to 1.
}
\examples{
\dontrun{
source_noaa <- system.file("extdata","signif.txt",package="NOAAsignifEarthQuakes",mustWork=TRUE)
noaa_raw <- load_NOAA_db(source_noaa)
noaa_raw \%>\%dplyr::select('DAY','MONTH','YEAR') \%>\% eq_build_date()
}
}
\seealso{
eq_clean_data
}
