% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{eq_build_location}
\alias{eq_build_location}
\title{Clean Location Names in NOAA data}
\usage{
eq_build_location(data)
}
\arguments{
\item{data}{raw NOAA data}
}
\value{
charater
}
\description{
Remove countries and unnecessary test,
clean spaces and switch to title cap.
}
\examples{
\dontrun{
source_noaa <- system.file("extdata","signif.txt",package="NOAAsignifEarthQuakes",mustWork=TRUE)
noaa_raw <- load_NOAA_db(source_noaa)
noaa_raw \%>\%dplyr::select('LOCATION_NAME') \%>\% eq_build_date()
}
}
\seealso{
eq_clean_data
}
