% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Perform the cleaning of the NOAA data}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{dbl_df dataframe input data frame}
}
\value{
dbl_df dataframe
}
\description{
Process the data to obtain the followning features:

* DATE: date of the earthquake

* LONGITUDE, LATITUDE: gps coordinates of the earthquake

* LOCATION_NAME: Location of the earthquake

* COUNTRY: Country where the earthquke occured

* DEATHS: Number of fatalities due to the earthquakes and related events

* MAG: Richter (equivalent) of the maginitude of the earthquake
}
\details{
Features YEAR, MONTH and DAY are selected to define the feature DATE
buld through function build_date.

Features LONGITUDE, LATITUDE are simply converted into numeric

Feature COUNTRY is left unchanged by the LOCATION_NAME is processed through package function clean_location.

Select features: LONGITUDE, LATITUDE, COUNTRY, LOCATION_NAME to define location.
LOCATION_NAME

Selected features TOTAL_DEATHS and EQ_RPIMARY are left unchanged but renames DEATHS and MAG
}
\examples{
\dontrun{
clean_data <- eq_clean_data(raw_data)
}
}
\seealso{
build_date clean_location
}
