% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_NOAA.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create html labels for Earthquke data}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{processed NOAA data}
}
\value{
character of the same lenght as the input data
}
\description{
Create popup content as a html encoded character.
Indicates location if value is not missing (NA),
richter scale magnitude if value is not missing(NA),
total number of deaths if value is not missing(NA).
}
\examples{
\dontrun{
noaa <- load_NOAA_db(file_noaa)\%>\% eq_clean_data()
noaa \%>\%eq_create_label()
noaa_with_label -> noaa \%>\%
    dplyr::mutate("popup_text"=eq_create_label(.))
}

}
