% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_NOAA_timeline.R
\name{eq_legend_timeline}
\alias{eq_legend_timeline}
\title{Transform NOAA feature into human readable character}
\usage{
eq_legend_timeline(quo)
}
\arguments{
\item{quo}{quosure corresponding to mapping of a variable}
}
\value{
Character
}
\description{
For DEATHS feature returns "Total nb of deaths",
for MAG feature returns "Richter scale magnitude"
and for any other return the feature name in title case.
}
\examples{
\dontrun{
 mapping <- ggplot2::aes(x=DATE,fill=MAG,size=DEATHS)
 eq_legend_timeline(mapping$x)
 eq_legend_timeline(mapping$fill)
 "Total nb of deaths"
 eq_legend_timeline(mapping$size)
}

}
