% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_NOAA_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Displays a timeline of Earthquakes}
\usage{
geom_timeline(data = NULL,mapping=NULL,xmin=NULL,xmax=NULL, countries= NULL,alpha=0.5)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{xmin}{start date for data vizualisation [Character] encoded as YYYY-MM-DD or [Date]}

\item{xmax}{end date for data vizualisation [Character] encoded as YYYY-MM-DD or [Date]}

\item{countries}{[Character] list of Countries to filter}

\item{alpha}{alpha parameter for transparency}
}
\value{
ggplot plot object
}
\description{
For a given (optional) selection of date range and countries.
The resulting selection of earthquakes is represented by a circle along a time axis.
Optional feature enable to add a vertical axis usualy country aesthetic \code{y}.
Additional optional aesthetics \code{fill} an \code{r} enable the user to represent
additional continuous variables such as nb of deaths or magnitude.
}
\examples{
\dontrun{
ggplot(data=clean_data,aes(x=DATE,y=COUNTRY)) + geom_timeline()
}

}
\seealso{
timeline_data geom_timeline_label relabel_legend
}
