% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_NOAA_timeline.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(data=NULL, n_max =5, ...)
}
\arguments{
\item{data}{input data transformed for vizualisation}

\item{n_max}{maximun number of earthquake to label by countries default value 5}

\item{...}{inherited parameters}
}
\value{
ggplot2 layer
}
\description{
geom_timeline_label
}
\examples{
\dontrun{
g <- ggplot(data=clean_data,aes(x=DATE,y=LOCATION_NAME)) +
     geom_timeline() +
     geom_timeline_label()
}
}
\seealso{
timeline_label geom_timeline
}
