% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize each word}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{input character string}
}
\value{
character string
}
\description{
Take a character string as input, split into words
the first letter of each word set to upper case next to lower case
}
