% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_NOAA_timeline.R
\name{timeline_data}
\alias{timeline_data}
\title{Filter NOAA processed data by date and countries}
\usage{
timeline_data(data=NULL,dmin=NULL,dmax=NULL,countries=NULL)
}
\arguments{
\item{data}{input tbl_df dataframe}

\item{dmin}{start date for data vizualisation [Character] encoded as YYYY-MM-DD or [Date]}

\item{dmax}{end date for data vizualisation [Character] encoded as YYYY-MM-DD or [Date]}

\item{countries}{[Character] list of Countries to filter}
}
\value{
tbl_df dataframe
}
\description{
Prepare data for visualisation, tking the clean data as input,
apply some filters in term of date range and countries in order
to limit the amount of earthquakes to display.
}
\examples{
\dontrun{
filtered_data <- timeline_data(clean_noaa,
                               dmin='2010-01-01',
                               dmax='2018-01-01',
                               countries=c('USA','China'))
}
}
\seealso{
timeline_draw geom_timeline
}
