% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_NOAA_timeline.R
\name{timeline_draw}
\alias{timeline_draw}
\title{Build a grob describing a timeline of earthquake events}
\usage{
timeline_draw(data, panel_scales,coord)
}
\arguments{
\item{data}{tbl_df input untransformed dataframe}

\item{panel_scales}{panel parameters}

\item{coord}{Coord object required to transform data coordinates to figure coordinates}
}
\value{
gTree instance
}
\description{
Each eathquakes is represented by a circle grob along a time axis.
Optional feature enable to add a vertical axis usualy country aesthetic \code{y}.
Additional optional aesthetics \code{fill} an \code{r} enable the user to represent
additional continuous variables such as nb of deaths or magnitude.
}
\examples{
\dontrun{
GeomTimeLine <- ggplot2:: ggproto(
     "GeomTimeLine", ggplot2::Geom,
     required_aes = c('x'),
     default_aes = ggplot2::aes(y=0.5,shape=21,size=1,fill=1,r=10),
     draw_key = c("r","fill"),
     draw_panel = timeline_draw
  )
ggplot2:: ggplot()+
ggplot2:: layer(
    geom = GeomTimeLine, mapping = ggplot2:: aes(x=DATE),
    data = data, stat = "identity", position = "identity",
    show.legend = TRUE, inherit.aes = TRUE,
    params = list(na.rm = TRUE)
  )
}
}
\seealso{
timeline_data geom_timeline
}
