% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_NOAA_timeline.R
\name{timeline_label}
\alias{timeline_label}
\title{timeline_label}
\usage{
timeline_label(data, panel_scales,coord,n_max=NULL)
}
\arguments{
\item{data}{input data transformed for vizualisation}

\item{panel_scales}{panel parameters}

\item{coord}{Coord object required to transform data coordinates to figure coordinates}

\item{n_max}{parameter number of earthquake location to display by country}
}
\value{
gTree instance
}
\description{
timeline_label
}
\seealso{
geom_timeline_label
}
