% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_NOAA_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Perform the cleaning of the NOAA data}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{dbl_df dataframe input data frame}
}
\value{
dbl_df dataframe
}
\description{
Process the data to obtain the followning features:

* DATE: date of the earthquake

* LONGITUDE, LATITUDE: gps coordinates of the earthquake

* LOCATION_NAME: Location of the earthquake

* COUNTRY: Country where the earthquke occured

* DEATHS: Number of fatalities due to the earthquakes and related events

* MAG: Richter (equivalent) of the maginitude of the earthquake
}
\details{
Features YEAR, MONTH and DAY are selected to define the feature DATE through function eq_build_date.
Feature LOCATION_NAME is processed through package function eq_build_location.

Select features: LONGITUDE, LATITUDE, COUNTRY, LOCATION_NAME to define location.
LOCATION_NAME

Selected features TOTAL_DEATHS and EQ_RPIMARY are left unchanged but renamed DEATHS and MAG
}
\examples{
\dontrun{
source_noaa <- system.file("extdata","signif.txt",package="NOAAsignifEarthQuakes",mustWork=TRUE)
noaa_raw <- load_NOAA_db(source_noaa)
noaa_raw \%>\% eq_clean_data()
}
}
\seealso{
eq_build_date eq_build_location
}
